catch most symlink loops quickly
With this optimization, a cleverly crafted giant nest of symlinks could
make a fork bomb run out of memory faster than it would without this
optimization.  The attacker could control the keys of an associative array.

Of course, most systems >:-} will have a GNU `readlink -f` so no need to
worry in real life.